endpoint: move most bpf, bpf map dependencies into bpf.go
The code dealing with bpf-related functionality should be
moved into a file that utilizes bpf-related code, bpf.go.

Signed-off by: Ian Vernon <ian@cilium.io>
